# specs/004-ai-todo-chatbot/contracts/chat.openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Todo Chat API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/chat/": {
      "post": {
        "summary": "Send a message to the AI agent",
        "tags": ["chat"],
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "type": "string",
                    "example": "Add a task to record a demo video"
                  },
                  "conversation_id": {
                    "type": "string",
                    "format": "uuid",
                    "description": "Optional: ID of existing conversation"
                  }
                },
                "required": ["message"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI response and updated conversation metadata",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "response": {
                      "type": "string",
                      "example": "I've added the task for you!"
                    },
                    "conversation_id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["success", "tool_call", "error"]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/history/{conversation_id}": {
      "get": {
        "summary": "Retrieve conversation history",
        "tags": ["chat"],
        "security": [{"bearerAuth": []}],
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "responses": {
          "200": {
            "description": "List of messages in the conversation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "role": {"type": "string"},
                      "content": {"type": "string"},
                      "created_at": {"type": "string", "format": "date-time"}
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}